+++ src/components/Input/InputText.module.scss
@@ -1,0 +1,30 @@
+@import 'assets/styles/global.scss';
+.inputContainer {
+  border-radius: 8px;
+  border: 1px solid var(--gray-300, #ccc);
+  background: var(--white, #fff);
+  width: 72rem;
+  height: 5rem;
+  padding: 1.2rem 1.6rem;
+  font-size: 1.6rem;
+  font-weight: 400;
+  @include mobile {
+    width: 32rem;
+  }
+}
+.inputContainer:focus {
+  border: 2px solid var(--gray-700, #3a3a3a);
+}
+.inputContainer:hover {
+  border: 1px solid var(--gray-500, #555);
+}
+.isInputValue {
+  border: 2px solid var(--gray-700, #3a3a3a);
+}
+.error {
+  border: 1px solid var(--Error, #dc3a3a);
+}
+.errorMsg {
+  color: red;
+  font-size: 1.6rem;
+}
